Now takes earliest time if dup
When practice mojo sees that the patient has two appts the same day it makes it into a family and puts the times not in descending order. The program now checks the time the patient comes in and takes the earliest time as that's when the pt will be arriving.