Comma change failed
Spaces werent being sent when inputing for the reminders so it would be "LastJr, First" instead of "Last Jr, First". Attempted to add space before caps and if line equals the double comma line. Need to still add in date change in softdent contact according to the date entered by user.